+++ R/utils.R
@@ -1,37 +1,3 @@
-now <- function(x = Sys.time(), format = "[%H:%M:%OS3] ") {
-  ## format(x, format = format) ## slower
-  format(as.POSIXlt(x, tz = ""), format = format)
-}
-
-mdebug <- function(..., prefix = now(), debug = isTRUE(getOption("future.mirai.debug"))) {
-  if (!debug) return()
-  message(prefix, ...)
-}
-
-mdebugf <- function(..., appendLF = TRUE,
-                    prefix = now(), debug = isTRUE(getOption("future.mirai.debug"))) {
-  if (!debug) return()
-  message(prefix, sprintf(...), appendLF = appendLF)
-}
-
-#' @importFrom utils capture.output str
-mstr <- function(..., prefix = now(), debug = isTRUE(getOption("future.mirai.debug"))) {
-  if (!debug) return()
-  stdout <- capture.output(str(...))
-  stdout <- paste(prefix, stdout, sep = "", collapse = "\n")
-  message(stdout)
-}
-
-#' @importFrom utils capture.output str
-mprint <- function(..., prefix = now(), debug = isTRUE(getOption("future.mirai.debug"))) {
-  if (!debug) return()
-  stdout <- capture.output(print(...))
-  stdout <- paste(prefix, stdout, sep = "", collapse = "\n")
-  message(stdout)
-}
-
-mprintf <- function(...) message(now(), sprintf(...), appendLF = FALSE)
-
 stop_if_not <- function(...) {
   res <- list(...)
   for (ii in 1L:length(res)) {