+++ inst/testme/_prologue/002.record-state.R
@@ -1,0 +1,4 @@
+## Record original state
+ovars <- ls(envir = globalenv())
+oenvs <- oenvs0 <- Sys.getenv()
+oopts0 <- options()