+++ inst/testme/_prologue/030.imports.R
@@ -1,0 +1,3 @@
+## Private future.mirai functions
+mprint <- future.mirai:::mprint
+mprintf <- future.mirai:::mprintf