+++ inst/testme/_prologue/050.utils.R
@@ -1,0 +1,43 @@
+attach_locally <- function(x, envir = parent.frame()) {
+  for (name in names(x)) {
+    assign(name, value = x[[name]], envir = envir)
+  }
+}
+
+
+## From R.utils 2.0.2 (2015-05-23)
+hpaste <- function(..., sep="", collapse=", ", last_collapse=NULL,
+                   max_head=if (missing(last_collapse)) 3 else Inf,
+                   max_tail=if (is.finite(max_head)) 1 else Inf,
+                   abbreviate="...") {
+  max_head <- as.double(max_head)
+  max_tail <- as.double(max_tail)
+  if (is.null(last_collapse)) last_collapse <- collapse
+
+  # Build vector 'x'
+  x <- paste(..., sep = sep)
+  n <- length(x)
+
+  # Nothing todo?
+  if (n == 0) return(x)
+  if (is.null(collapse)) return(x)
+
+  # Abbreviate?
+  if (n > max_head + max_tail + 1) {
+    head <- x[seq_len(max_head)]
+    tail <- rev(rev(x)[seq_len(max_tail)])
+    x <- c(head, abbreviate, tail)
+    n <- length(x)
+  }
+
+  if (!is.null(collapse) && n > 1) {
+    if (last_collapse == collapse) {
+      x <- paste(x, collapse = collapse)
+    } else {
+      x_head <- paste(x[1:(n - 1)], collapse = collapse)
+      x <- paste(x_head, x[n], sep = last_collapse)
+    }
+  }
+
+  x
+}