+++ inst/testme/_prologue/090.options.R
@@ -1,0 +1,12 @@
+## Default options
+oopts <- options(
+  warn = 1L,
+  showNCalls = 500L,
+  mc.cores = 2L,
+
+future.debug = FALSE,
+  ## Reset the following during testing in case
+  ## they are set on the test system
+  future.availableCores.system = NULL,
+  future.availableCores.fallback = NULL
+)