+++ tests/incl/start.R
@@ -1,2 +1,0 @@
-library("future.mirai")
-source("incl/start,load-only.R")