+++ NEWS.md
@@ -1,7 +1,9 @@
 # Version (development version)
 
- * ...
- 
+## New Features
+
+ * Now 'mirai' futures can be interrupted using `interrupt()`.
+
 
 # Version 0.2.2